Moved pins to core module
