+++ core/templates/core/pins/default.html
