--- user_interaction/templates/user_interaction/snippets/user_pins.html
@@ -1,4 +1,0 @@
-
-{% for pintype in pins %}
-    {% include pintype.template with pintype=pintype %}
-{% endfor %}
